remove serialization tests
