+++ src/NMoneys.Tests/NMoneys.Tests.csproj
@@ -53,6 +53,7 @@
     </Reference>
   </ItemGroup>
   <ItemGroup>
+    <Compile Include="CustomConstraints\Must.Extensions.net.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="..\NMoneys\Properties\NMoneys_AssemblyInfo.cs">
       <Link>Properties\NMoneys_AssemblyInfo.cs</Link>
@@ -64,17 +65,17 @@
     <Compile Include="Allocations\RatioTester.cs" />
     <Compile Include="Allocations\RatioTester.Equality.cs" />
     <Compile Include="CharacterReferenceTester.cs" />
-    <Compile Include="CurrencyIsoCodeTester.Serialization.cs" />
-    <Compile Include="CurrencyIsoCodeTester.Serialization.Xml.cs" />
-    <Compile Include="CurrencyTester.Serialization.Xml.cs" />
+    <Compile Include="CurrencyIsoCodeTester.Serialization.net.cs" />
+    <Compile Include="CurrencyIsoCodeTester.Serialization.Xml.net.cs" />
+    <Compile Include="CurrencyTester.Serialization.Xml.net.cs" />
     <Compile Include="CurrencyTester.Serialization.net.cs" />
     <Compile Include="CurrencyTester.Markup.cs" />
     <Compile Include="CurrencyTester.Code.cs" />
     <Compile Include="CurrencyTester.Creation.cs" />
     <Compile Include="CurrencyTester.Comparisons.cs" />
     <Compile Include="MoneyTester.Creation.cs" />
-    <Compile Include="MoneyTester.Serialization.Xml.cs" />
-    <Compile Include="MoneyTester.Serialization.cs" />
+    <Compile Include="MoneyTester.Serialization.Xml.net.cs" />
+    <Compile Include="MoneyTester.Serialization.net.cs" />
     <Compile Include="MoneyTester.Arithmetic.cs" />
     <Compile Include="MoneyTester.Arithmetic.Operators.cs" />
     <Compile Include="MoneyTester.Parsing.cs" />
